chore(r/adbcpostgresql): Add windows build support (#583)
Following the approach here: https://github.com/r-dbi/RPostgres .
Thankfully, there are static builds of libpq (albeit an old ish version)
available via rwinlibs so we don't have to vendor.

There were a few changes required in the driver. First, there is no
`nthll()` in the gcc headers in mingw32/64, so we need a home-grown
bswap64. I took the approach that S2 did (
https://github.com/google/s2geometry/blob/master/src/s2/base/port.h#L236
). Second, the toolchain for R 3.6 on Windows is gcc 4.8, which doesn't
support enums as `unordered_map<>` keys. One could also conditionally
define a custom hasher but it seemed like a static cast to int wasn't
too intrusive.

+++ c/driver/postgresql/type.cc
@@ -29,21 +29,21 @@ void TypeMapping::Insert(uint32_t oid, const char* typname, const char* typrecei
   // Record 'canonical' types
   if (std::strcmp(typname, "int4") == 0) {
     // DCHECK_EQ(type, PgType::kInt4);
-    canonical_types[PgType::kInt4] = oid;
+    canonical_types[static_cast<int32_t>(PgType::kInt4)] = oid;
   } else if (std::strcmp(typname, "int8") == 0) {
     // DCHECK_EQ(type, PgType::kInt8);
-    canonical_types[PgType::kInt8] = oid;
+    canonical_types[static_cast<int32_t>(PgType::kInt8)] = oid;
   } else if (std::strcmp(typname, "float8") == 0) {
     // DCHECK_EQ(type, PgType::kFloat8);
-    canonical_types[PgType::kFloat8] = oid;
+    canonical_types[static_cast<int32_t>(PgType::kFloat8)] = oid;
   } else if (std::strcmp(typname, "text") == 0) {
-    canonical_types[PgType::kText] = oid;
+    canonical_types[static_cast<int32_t>(PgType::kText)] = oid;
   }
   // TODO: fill in remainder
 }
 
 uint32_t TypeMapping::GetOid(PgType type) const {
-  auto it = canonical_types.find(type);
+  auto it = canonical_types.find(static_cast<int32_t>(type));
   if (it == canonical_types.end()) {
     return 0;
   }
